pyproject.toml
[tool.poetry]
name = "eurostable"
version = "0.1.0"
description = ""
authors = ["Legacy Coding Studios"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.17.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
